整体测试，后台防伪码查询，验证积分配置

+++ src/com/sys/dao/ScoreDao.java
@@ -43,24 +43,24 @@ public class ScoreDao extends HibernateDao<Score> {
 		return Double.valueOf(query.uniqueResult().toString());
 	}
 
-	public int getFactoryHistoryScore(String factoryId) {
+	public double getFactoryHistoryScore(String factoryId) {
 		StringBuffer sql = new StringBuffer(" select coalesce(sum(s.score),0) " +
 			"from t_score s where s.factory_id =:factoryId and score>=0");
 		
 		Query query = this.getSession().createSQLQuery(sql.toString()).
 		setParameter("factoryId", factoryId);
 
-		return Integer.valueOf(query.uniqueResult().toString());
+		return Double.valueOf(query.uniqueResult().toString());
 	}
 	
-	public int getFactoryConsumeScore(String factoryId) {
+	public double getFactoryConsumeScore(String factoryId) {
 		StringBuffer sql = new StringBuffer(" select coalesce(sum(s.score),0) " +
 			"from t_score s where s.factory_id =:factoryId and score <0");
 		
 		Query query = this.getSession().createSQLQuery(sql.toString()).
 		setParameter("factoryId", factoryId);
 
-		return Integer.valueOf(query.uniqueResult().toString());
+		return Double.valueOf(query.uniqueResult().toString());
 	}
 	
 }
